requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.3.2
constantly==15.1.0
cryptography==42.0.2
daphne==4.0.0
Django==3.2.23
django-cors-headers==4.1.0
django-environ==0.11.2
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==6.7.0
incremental==22.10.0
inflection==0.5.1
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.1
selenium==4.11.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
Twisted==22.4.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.6
wsproto==1.2.0
zipp==3.15.0
zope.interface==6.1
